+++ src/Sienar.Plugin.Cms.Client/_Imports.razor
@@ -16,6 +16,7 @@
 @using Sienar.Services
 @using Sienar.Layouts
 @using Sienar.Pages
+@using Sienar.Security
 @using Sienar.Ui
 
 @attribute [Authorize]
