+++ src/Sienar.Plugin.Cms.Core/Identity/Requests/LoginRequest.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using Sienar.Data;
+using Sienar.Security;
 using Sienar.Services;
 
 namespace Sienar.Identity.Requests;
