+++ src/Sienar.Plugin.Cms.Core/Identity/Requests/RegisterRequest.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using Sienar.Validation;
-using Sienar.Data;
+using Sienar.Security;
 using Sienar.Services;
 
 namespace Sienar.Identity.Requests;
