+++ src/Sienar.Plugin.Cms.Core/Identity/Requests/ResetPasswordRequest.cs
@@ -2,7 +2,7 @@ using System;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using Sienar.Validation;
-using Sienar.Data;
+using Sienar.Security;
 using Sienar.Services;
 
 namespace Sienar.Identity.Requests;
