+++ src/Sienar.Plugin.Cms.Server/Hooks/UserIsAdminAccessValidator.cs
@@ -1,4 +1,5 @@
 ﻿using Sienar.Infrastructure;
+using Sienar.Security;
 
 namespace Sienar.Hooks;
