+++ src/Sienar.Plugin.Cms.Server/Identity/Processors/DeleteAccountProcessor.cs
@@ -3,10 +3,10 @@
 using System.Threading.Tasks;
 using Sienar.Errors;
 using Sienar.Identity.Requests;
-using Sienar.Infrastructure;
 using Sienar.Data;
 using Sienar.Identity.Data;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
