+++ src/Sienar.Plugin.Cms.Server/Identity/Processors/GetAccountDataProcessor.cs
@@ -5,8 +5,8 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using Sienar.Data;
 using Sienar.Identity.Results;
-using Sienar.Infrastructure;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
