+++ src/Sienar.Plugin.Cms.Server/Identity/Processors/InitiateEmailChangeProcessor.cs
@@ -7,10 +7,10 @@ using Sienar.Email;
 using Sienar.Errors;
 using Sienar.Extensions;
 using Sienar.Identity.Requests;
-using Sienar.Infrastructure;
 using Sienar.Data;
 using Sienar.Identity.Data;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
