+++ src/Sienar.Plugin.Cms.Server/Identity/Processors/PerformEmailChangeProcessor.cs
@@ -6,10 +6,10 @@ using Sienar.Configuration;
 using Sienar.Email;
 using Sienar.Errors;
 using Sienar.Identity.Requests;
-using Sienar.Infrastructure;
 using Sienar.Data;
 using Sienar.Identity.Data;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
