+++ src/Sienar.Plugin.Cms.Server/Media/Hooks/VerifyUserCanModifyFileHook.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Sienar.Infrastructure;
 using Sienar.Hooks;
+using Sienar.Security;
 
 namespace Sienar.Media.Hooks;
