+++ src/Sienar.Plugin.Cms.Server/Security/HttpContextUserAccessor.cs
@@ -4,7 +4,7 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 
-namespace Sienar.Infrastructure;
+namespace Sienar.Security;
 
 public class HttpContextUserAccessor : IUserAccessor
 {
