+++ src/Sienar.Utils/Services/Service.cs
@@ -7,6 +7,7 @@ using Sienar.Data;
 using Sienar.Hooks;
 using Sienar.Infrastructure;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Services;
 
@@ -17,7 +18,7 @@ public class Service<TRequest, TResult> : ServiceBase, IService<TRequest, TResul
 {
 	private readonly ILogger<Service<TRequest, TResult>> _logger;
 	private readonly IBotDetector _botDetector;
-	private readonly IAccessValidatorService<TRequest> _accessValidator;
+	private readonly IAccessValidationRunner<TRequest> _accessValidator;
 	private readonly IStateValidatorService<TRequest> _stateValidator;
 	private readonly IBeforeActionService<TRequest> _beforeHooks;
 	private readonly IAfterActionService<TRequest> _afterHooks;
@@ -26,7 +27,7 @@ public class Service<TRequest, TResult> : ServiceBase, IService<TRequest, TResul
 	public Service(
 		ILogger<Service<TRequest, TResult>> logger,
 		IBotDetector botDetector,
-		IAccessValidatorService<TRequest> accessValidator,
+		IAccessValidationRunner<TRequest> accessValidator,
 		IStateValidatorService<TRequest> stateValidator,
 		IBeforeActionService<TRequest> beforeHooks,
 		IAfterActionService<TRequest> afterHooks,
