+++ src/Sienar.Utils/Ui/Username.razor
@@ -1,4 +1,5 @@
-﻿@implements IDisposable
+﻿@using Sienar.Security
+@implements IDisposable
 @inject AuthenticationStateProvider AuthStateProvider
 @inject IUserAccessor UserAccessor
